use splitchunks for vendor bundling
instead of defining all types of imports in a seperate bundle.js file
using the webpack splitchunks optimization plugin. The cacheGroup as it
is defined now, will bundle all libraries from the node_modules directory
into a file called bundle.

As a result we don't need to define a seperate bundle entry anymore.
In welcome.jsp, import the welcome.js as a module. This allows to use the import statement in the welcome.js file.

In welcome.js added the required import statements that are used by this specific page. These imports are bundled into the bundle.js.

removed obsolete files which are no longer needed.